Added GitHubServiceNoop for integration tests
GtiHubService can be disabled by setting "github.enabled=false". In that case a no-op service
is used instead.
