+++ runtime/src/main/resources/application.yml
@@ -62,3 +62,7 @@ endpoints:
 dao:
   kind: jsondb
 
+github:
+  enabled: true
+  service: ipaas-github-proxy
+
