+++ runtime/src/test/resources/application-test.yml
@@ -24,3 +24,6 @@ spring:
     username: sa
     password: sa
     driver-class-name: org.h2.Driver
+
+github:
+  enabled: false
